Rename libPlatform to libStatistics
