+++ src/Spectre.libStatistics.Tests/Spectre.libStatistics.Tests.vcxproj
@@ -13,7 +13,7 @@
   <PropertyGroup Label="Globals">
     <VCProjectVersion>15.0</VCProjectVersion>
     <ProjectGuid>{81F10DE8-B9EF-400F-B3A7-57F2AFF35A14}</ProjectGuid>
-    <RootNamespace>SpectrelibPlatformTests</RootNamespace>
+    <RootNamespace>SpectrelibStatisticsTests</RootNamespace>
     <WindowsTargetPlatformVersion>8.1</WindowsTargetPlatformVersion>
   </PropertyGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.Default.props" />
