+++ src/Spectre.libStatistics.Tests/packages.config
